DT: soc: t186: pwm: add clock parent for PWM to PLLP_OUT0
For PWM1..3 and PWM5..8, the allowed parents are CLK_M and PLLP_OUT0.
For PWM4, it is oscillator clocks.

Add DT property for parent clock under all PWM nodes except PWM4.
This will provide the PWM clock source to 102MHz Maximum and PWM
output can go to 396KHz.

Make this change for CVM only and still keeping the default parent
for VCM. This is because some of test tools need to be updated.
Once tools get updated, it will also be enabled for VCM.

Bug 200267484
Bug 200299608

Change-Id: I9d80073a76273fdae98f471d04a09d27c1bf523e
Signed-off-by: R Raj Kumar <rrajk@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1467617
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit